fix extended contract flow

+++ purchasing/data/contracts.py
@@ -120,10 +120,9 @@ def extend_a_contract(child_contract_id=None, delete_child=True):
 
     # strip the flow -- this will serve as the flag that
     # the contract is extended
-    parent_contract.flow = None
-    parent_contract.flow_id = None
 
-    delete_contract(child_contract_id)
+    if delete_child:
+        delete_contract(child_contract_id)
 
     return parent_contract
